+++ app/assets/javascripts/objects/reports.js
@@ -9,8 +9,19 @@ function reportRowsSort() {
   });
 }
 
+function toggleReportFields(value) {
+  $("[data-report-types]").hide();
+  if (value != "") {
+    $("[data-report-types~=" + value + "]").show();
+  }
+}
+
 $(document)
   .on("click", "[data-object~=remove-report-row]", function() {
     $(this).closest("[data-object~=report-row]").remove();
+    return false;
+  })
+  .on("change", "#report_report_type", function() {
+    toggleReportFields($(this).val());
     return false;
   });