+++ test/controllers/reports_controller_test.rb
@@ -13,6 +13,7 @@ class ReportsControllerTest < ActionDispatch::IntegrationTest
   def report_params
     {
       project_id: projects(:one).id,
+      report_type: "expressions_by_site",
       name: "New Report",
       header_label: "Diagnosis",
       archived: "0",
@@ -51,9 +52,15 @@ class ReportsControllerTest < ActionDispatch::IntegrationTest
     assert_redirected_to report_url(Report.last)
   end
 
-  test "should show report" do
+  test "should show expressions by site report" do
     login(@admin)
     get report_url(@report)
+    assert_response :success
+  end
+
+  test "should show randomizations by site report" do
+    login(@admin)
+    get report_url(reports(:randomizations))
     assert_response :success
   end